pyproject.toml
[tool.poetry]
name = "pokestar"
version = "0.1.0"
description = ""
authors = ["Gonzalo Sanchez"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.30.0"
urllib3 = "1.26.15"
django = "^4.2.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "^8.13.2"
pre-commit = "^3.3.2"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
color = true

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
